refactor(telemetry): remove instrumentation_stalled, keep the clock it hid
The event claimed to turn an uninterpretable absence into a signal and did not. It fired
only on a flush tick or a graceful shutdown, so a SIGKILLed daemon sent daemon_started
and nothing else and stayed in the absence regardless. Everything it carried is
derivable from daemon_started minus app_instrumented on sessionId, and that set
difference is strictly more complete because it catches the killed daemons. Two ways to
compute one number that disagree is worse than one way.

Two defects leave with it: the shutdown path had no floor, so a run cancelled after
three seconds in the wrong directory emitted a stall and inflated every historical
count by an unknown amount; and four different wiring failures shared one bucket.

The CLOCK is not the event and stays. #425 wired two user-facing features to it hours
before this removal was written up: the no-session diagnosis reports how long this daemon
has waited, and the flush path warns on stderr past ten minutes with nothing connected.
Deleting the module as scoped would have taken both with it. The survivors moved to
session/stall-clock.ts, beside their readers, since living under telemetry/ is what made
them look like part of the metric.

The stderr warning is now driven by the clock directly and keeps its once-per-run guard.
The shutdown emission is gone entirely rather than rewired, because it existed only to
feed the metric and carried the floor-less bug.
