+++ packages/server/src/telemetry/daemon-telemetry.ts
@@ -13,12 +13,7 @@ import { profileProject } from './project-profile.js';
 import { startUpdateCheck } from '../update/update-nudge.js';
 import { markDaemonStart } from './mcp-connection.js';
 import { markInstrumentationClock } from './app-instrumented.js';
-import {
-  markStallClock,
-  reportInstrumentationStall,
-  STALL_AFTER_MS,
-} from './instrumentation-stall.js';
-import { readProjectId } from '../cli/cli-port.js';
+import { markStallClock, STALL_AFTER_MS, stallUptime } from '../session/stall-clock.js';
 
 /** Let the daemon finish coming up before walking the source tree for a profile. */
 const PROJECT_PROFILE_DELAY_MS = 5_000;
@@ -100,22 +95,19 @@ export function installDaemonTelemetry(
     }
   }, PROJECT_PROFILE_DELAY_MS).unref();
 
+  // Once per daemon run: a warning repeated every ten minutes at somebody already stuck is noise.
+  let stallWarned = false;
+
   const flushWindow = (): void => {
-    // BEFORE the idle return, and that ordering is the whole point: a stalled install produces an
-    // empty window by definition — no app, so no tool calls, so nothing to roll up. Checking it
-    // after the early return would make the one case this event exists for the one case it never
-    // sees. It is self-limiting (once per run, never after an app connects), so running it on every
-    // tick costs a comparison.
-    const stalled = reportInstrumentationStall(
-      {
-        // Same test `app_instrumented` uses for the same field, so the success and failure events
-        // stay directly comparable: a stamped projectId means `init` has run here.
-        initialized: readProjectId(cwd) !== undefined,
-        agentAttached: metrics.agentEverAttached,
-      },
-      now,
-    );
-    if (stalled) {
+    // BEFORE the idle return, and that ordering still matters for the reason the removed event
+    // documented: a stalled install produces an empty window by definition — no app, so no tool
+    // calls, so nothing to roll up — so checking after the early return would never fire on the one
+    // case worth warning about. What is left here is the WARNING, for a human watching the log; the
+    // metric that used to ride alongside it was removed as a second, less complete way to compute
+    // `daemon_started` minus `app_instrumented`.
+    const waited = stallUptime(now());
+    if (!stallWarned && waited !== undefined && waited >= STALL_AFTER_MS) {
+      stallWarned = true;
       const minutes = Math.round(STALL_AFTER_MS / 60_000);
       process.stderr.write(
         `[reticle] no app has connected after ${String(minutes)} minutes — ` +
@@ -144,19 +136,6 @@ export function installDaemonTelemetry(
       stopped ??= (async () => {
         clearInterval(flush);
         clearTimeout(firstFlush);
-        // Last chance to say that no app ever arrived. The periodic path cannot report an
-        // UNATTENDED stall at all — that daemon idle-exits at five minutes and the periodic
-        // threshold is ten — so without this the event only ever describes daemons with an agent
-        // attached, which is not the population it was built to measure. Self-limiting: if the
-        // interval already reported, this returns false and emits nothing.
-        reportInstrumentationStall(
-          {
-            initialized: readProjectId(cwd) !== undefined,
-            agentAttached: metrics.agentEverAttached,
-          },
-          now,
-          { atShutdown: true },
-        );
         // The rich one: the whole session in a single event — duration, the tool histogram, error
         // shapes, verifications, browser launches. This is what replaced the per-tool-call event.
         await getTelemetry().emit(TelemetryEventKind.DAEMON_STOPPED, {
